+++ src/context.rs
@@ -138,7 +138,7 @@ where
     /// }
     ///
     /// # fn main() {
-    /// # let mut sys = System::new("test");
+    /// # let mut sys = System::new();
     /// let addr = sys.block_on(async { MyActor.start() });
     /// sys.run();
     /// # }