+++ src/stream.rs
@@ -72,7 +72,7 @@ where
     /// }
     ///
     /// fn main() {
-    ///     let mut sys = System::new("example");
+    ///     let mut sys = System::new();
     ///     let addr = sys.block_on(async { MyActor.start() });
     ///     sys.run();
     /// }