+++ tests/test_context.rs
@@ -95,54 +95,49 @@ impl Handler<TimeoutMessage> for MyActor {
 
 #[test]
 fn test_add_timeout() {
-    System::run(|| {
+    System::new().block_on(async {
         let _addr = MyActor { op: Op::Timeout }.start();
-    })
-    .unwrap();
+    });
 }
 
 #[test]
 fn test_add_timeout_cancel() {
-    System::run(|| {
+    System::new().block_on(async {
         let _addr = MyActor { op: Op::Cancel }.start();
 
         actix_rt::spawn(async move {
             sleep(Duration::new(0, 1000)).await;
             System::current().stop();
         });
-    })
-    .unwrap();
+    });
 }
 
 #[test]
 // delayed notification should be dropped after context stop
 fn test_add_timeout_stop() {
-    System::run(|| {
+    System::new().block_on(async {
         let _addr = MyActor {
             op: Op::TimeoutStop,
         }
         .start();
-    })
-    .unwrap();
+    });
 }
 
 #[test]
 fn test_run_after() {
-    System::run(|| {
+    System::new().block_on(async {
         let _addr = MyActor { op: Op::RunAfter }.start();
-    })
-    .unwrap();
+    });
 }
 
 #[test]
 fn test_run_after_stop() {
-    System::run(|| {
+    System::new().block_on(async {
         let _addr = MyActor {
             op: Op::RunAfterStop,
         }
         .start();
-    })
-    .unwrap();
+    });
 }
 
 struct ContextWait {
@@ -178,13 +173,15 @@ fn test_wait_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let cnt = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let addr = ContextWait { cnt }.start();
         addr.do_send(Ping);
         addr.do_send(Ping);
         addr.do_send(Ping);
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m.load(Ordering::Relaxed), 1);
 }
@@ -194,7 +191,8 @@ fn test_message_stream_wait_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let _addr = ContextWait::create(move |ctx| {
             let (tx, rx) = unbounded_channel();
             let _ = tx.send(Ping);
@@ -205,8 +203,9 @@ fn test_message_stream_wait_context() {
             ctx.add_message_stream(StreamRx { rx });
             actor
         });
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m.load(Ordering::Relaxed), 1);
 }
@@ -216,7 +215,8 @@ fn test_stream_wait_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let _addr = ContextWait::create(move |ctx| {
             let (tx, rx) = unbounded_channel();
             let _ = tx.send(Ping);
@@ -226,8 +226,9 @@ fn test_stream_wait_context() {
             ctx.add_message_stream(StreamRx { rx });
             actor
         });
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m.load(Ordering::Relaxed), 1);
 }
@@ -249,7 +250,7 @@ impl Handler<Ping> for ContextNoWait {
     }
 }
 
-#[actix_rt::test]
+#[actix::test]
 async fn test_nowait_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let cnt = Arc::clone(&m);
@@ -266,7 +267,7 @@ async fn test_nowait_context() {
     assert_eq!(m.load(Ordering::Relaxed), 3);
 }
 
-#[actix_rt::test]
+#[actix::test]
 async fn test_message_stream_nowait_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
@@ -293,7 +294,8 @@ fn test_stream_nowait_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let _addr = ContextNoWait::create(move |ctx| {
             let (tx, rx) = unbounded_channel();
             let _ = tx.send(Ping);
@@ -308,8 +310,9 @@ fn test_stream_nowait_context() {
             sleep(Duration::from_millis(200)).await;
             System::current().stop();
         });
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m.load(Ordering::Relaxed), 3);
 }
@@ -319,7 +322,8 @@ fn test_notify() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let addr = ContextNoWait::create(move |ctx| {
             ctx.notify(Ping);
             ctx.notify(Ping);
@@ -333,8 +337,9 @@ fn test_notify() {
             sleep(Duration::from_millis(200)).await;
             System::current().stop();
         });
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m2.load(Ordering::Relaxed), 3);
 }
@@ -360,7 +365,8 @@ fn test_current_context_handle() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let _addr = ContextHandle::create(move |ctx| {
             h2.store(
                 ContextHandle::add_stream(once(async { Ping }), ctx).into_usize(),
@@ -369,8 +375,9 @@ fn test_current_context_handle() {
 
             ContextHandle { h: m2 }
         });
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m.load(Ordering::Relaxed), h.load(Ordering::Relaxed));
 }
@@ -382,7 +389,8 @@ fn test_start_from_context() {
     let m = Arc::new(AtomicUsize::new(0));
     let m2 = Arc::clone(&m);
 
-    System::run(move || {
+    let sys = System::new();
+    sys.block_on(async move {
         let _addr = ContextHandle::create(move |ctx| {
             h2.store(
                 ctx.add_stream(once(async { Ping })).into_usize(),
@@ -390,8 +398,9 @@ fn test_start_from_context() {
             );
             ContextHandle { h: m2 }
         });
-    })
-    .unwrap();
+    });
+
+    sys.run().unwrap();
 
     assert_eq!(m.load(Ordering::Relaxed), h.load(Ordering::Relaxed));
 }
@@ -418,7 +427,7 @@ impl StreamHandler<CancelPacket> for CancelHandler {
 
 #[test]
 fn test_cancel_handler() {
-    actix::System::run(|| {
+    actix::System::new().block_on(async {
         struct WtfStream {
             interval: tokio::time::Interval,
         }
@@ -442,8 +451,7 @@ fn test_cancel_handler() {
                 interval: interval_at(Instant::now(), Duration::from_millis(1)),
             }),
         });
-    })
-    .unwrap();
+    });
 }
 
 struct CancelLater {
@@ -475,7 +483,7 @@ impl Handler<CancelMessage> for CancelLater {
 
 #[test]
 fn test_cancel_completed_with_no_context_item() {
-    actix::System::run(|| {
+    actix::System::new().block_on(async {
         // first, spawn future that will complete immediately
         let addr = CancelLater { handle: None }.start();
 
@@ -491,6 +499,5 @@ fn test_cancel_completed_with_no_context_item() {
             sleep(Duration::from_millis(200)).await;
             System::current().stop();
         });
-    })
-    .unwrap();
+    });
 }